requirements.txt
streamlit
boto3
python-dotenv
